+++ src/utils/drag.js
@@ -2,13 +2,21 @@ const mousePosition = (e) => {
   e = e || window.event;
   const xPos = e.pageX
     ? e.pageX
-    : e.clientX + document.body.scrollLeft - document.body.clientLeft;
+    : (e.clientX + document.body.scrollLeft - document.body.clientLeft)
+    ? (e.clientX + document.body.scrollLeft - document.body.clientLeft)
+    : e.touches[0]
+    ? e.touches[0].clientX
+    : e.changedTouches[0].clientX;
   const yPos = e.pageY
     ? e.pageY
-    : e.clientY + document.body.scrollTop - document.body.clientTop;
+    : (e.clientY + document.body.scrollTop - document.body.clientTop)
+    ? (e.clientY + document.body.scrollTop - document.body.clientTop)
+    : e.touches[0]
+    ? e.touches[0].clientY
+    : e.changedTouches[0].clientY;
   return {
-    x: xPos,
-    y: yPos
+    x: xPos + Math.max(document.documentElement.scrollLeft, document.body.scrollLeft),
+    y: yPos + Math.max(document.documentElement.scrollTop, document.body.scrollTop)
   };
 };