+++ CHANGELOG.md
@@ -1,12 +1,24 @@
 # CHANGELOG
 
+### v2.2.0
+
+#### new props
+
+- Add `minuteStep` props
+- Add `limitDrag` props
+
+### v2.1.3
+
+- Bugfixed for drag position offset
+- Add `onTimezoneChange` callback
+
 ### v2.1.0
 
 #### new props
 
-- `phrases`: `React.PropTypes.object`
-- `timezone`: `React.PropTypes.string`
-- `onTimezoneChange`: `React.PropTypes.func`
+- `phrases`: `PropTypes.object`
+- `timezone`: `PropTypes.string`
+- `onTimezoneChange`: `PropTypes.func`
 
 ### v2.0.0
 
@@ -18,5 +30,5 @@
 
 #### new props
 
-- `showTimezone`: `React.PropTypes.bool`, default `false`
-- `timezone`:  `React.PropTypes.string`, default user current local timezone
+- `showTimezone`: `PropTypes.bool`, default `false`
+- `timezone`:  `PropTypes.string`, default user current local timezone